+++ src/DependencyInjection/NowoIconSelectorExtension.php
@@ -53,6 +53,7 @@ final class NowoIconSelectorExtension extends Extension implements PrependExtens
         $container->setParameter(Configuration::ALIAS . '.use_iconify_collection', $config['use_iconify_collection'] ?? false);
         $container->setParameter(Configuration::ALIAS . '.icons_api_path', $config['icons_api_path'] ?? '/api/icon-selector/icons');
         $container->setParameter(Configuration::ALIAS . '.form_theme', $config['form_theme'] ?? 'form_div_layout.html.twig');
+        $container->setParameter(Configuration::ALIAS . '.debug', $config['debug'] ?? false);
 
         if (!empty($config['use_iconify_collection'])) {
             $loader->load('services_iconify.yaml');